+++ packages/expo-media-library/android/src/main/java/expo/modules/medialibrary/next/objects/asset/delegates/AssetModernDelegate.kt
@@ -22,8 +22,10 @@ import expo.modules.medialibrary.next.extensions.resolver.queryAssetCreationTime
 import expo.modules.medialibrary.next.extensions.resolver.updateRelativePath
 import expo.modules.medialibrary.next.objects.wrappers.RelativePath
 import expo.modules.medialibrary.next.objects.asset.Asset
+import expo.modules.medialibrary.next.objects.asset.deleters.AssetDeleter
 import expo.modules.medialibrary.next.objects.wrappers.MediaType
 import expo.modules.medialibrary.next.objects.wrappers.MimeType
+import expo.modules.medialibrary.next.permissions.MediaStorePermissionsDelegate
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import java.io.File
@@ -32,7 +34,11 @@ import kotlin.time.DurationUnit
 import kotlin.time.toDuration
 
 @RequiresApi(Build.VERSION_CODES.Q)
-class AssetModernDelegate(override val contentUri: Uri, context: Context) : AssetDelegate {
+class AssetModernDelegate(
+  override val contentUri: Uri,
+  val assetDeleter: AssetDeleter,
+  context: Context
+): AssetDelegate {
   private val contextRef = WeakReference(context)
 
   private val contentResolver
@@ -109,18 +115,23 @@ class AssetModernDelegate(override val contentUri: Uri, context: Context) : Asse
       ?: MimeType.from(getUri())
   }
 
-  override suspend fun delete() {
-    contentResolver.delete(contentUri, null, null)
+  override suspend fun delete() = withContext(Dispatchers.IO) {
+    assetDeleter.delete(contentUri)
   }
 
   override suspend fun move(relativePath: RelativePath) {
     contentResolver.updateRelativePath(contentUri, relativePath)
+//    val newAssetUri = contentResolver.insertPendingAsset(getFilename(), getMimeType(), relativePath)
+//    assetDeleter.delete(contentUri)
+//    contentResolver.copyUriContent(contentUri, newAssetUri)
+//    contentResolver.publishPendingAsset(newAssetUri)
   }
 
   override suspend fun copy(relativePath: RelativePath): Asset = withContext(Dispatchers.IO) {
     val newAssetUri = contentResolver.insertPendingAsset(getFilename(), getMimeType(), relativePath)
     contentResolver.copyUriContent(contentUri, newAssetUri)
     contentResolver.publishPendingAsset(newAssetUri)
-    return@withContext Asset(newAssetUri, contextRef.getOrThrow())
+    val newAssetDelegate = AssetModernDelegate(newAssetUri, assetDeleter, contextRef.getOrThrow())
+    return@withContext Asset(newAssetDelegate)
   }
 }