+++ packages/expo-media-library/android/src/main/java/expo/modules/medialibrary/next/objects/asset/factories/AssetFactory.kt
@@ -3,7 +3,9 @@ package expo.modules.medialibrary.next.objects.asset.factories
 import android.net.Uri
 import expo.modules.medialibrary.next.objects.wrappers.RelativePath
 import expo.modules.medialibrary.next.objects.asset.Asset
+import expo.modules.medialibrary.next.objects.asset.delegates.AssetDelegate
 
-fun interface AssetFactory {
+interface AssetFactory {
+  fun create(contentUri: Uri): Asset
   suspend fun create(filePath: Uri, relativePath: RelativePath?): Asset
 }